reduce diffs
