+++ src/confluent_kafka/schema_registry/__init__.py
@@ -15,6 +15,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
+from typing import Optional
 
 from .schema_registry_client import (
   ConfigCompatibilityLevel,
@@ -34,16 +35,9 @@ from .schema_registry_client import (
   ServerConfig
 )
 
-from .common import (
-    _MAGIC_BYTE,
-    topic_subject_name_strategy,
-    topic_record_subject_name_strategy,
-    record_subject_name_strategy,
-    reference_subject_name_strategy
-)
+_MAGIC_BYTE = 0
 
 __all__ = [
-  "_MAGIC_BYTE",
   "ConfigCompatibilityLevel",
   "Metadata",
   "MetadataProperties",
@@ -61,6 +55,61 @@ __all__ = [
   "ServerConfig",
   "topic_subject_name_strategy",
   "topic_record_subject_name_strategy",
-  "record_subject_name_strategy",
-  "reference_subject_name_strategy",
+  "record_subject_name_strategy"
 ]
+
+
+def topic_subject_name_strategy(ctx, record_name: Optional[str]) -> Optional[str]:
+    """
+    Constructs a subject name in the form of {topic}-key|value.
+
+    Args:
+        ctx (SerializationContext): Metadata pertaining to the serialization
+            operation.
+
+        record_name (Optional[str]): Record name.
+
+    """
+    return ctx.topic + "-" + ctx.field
+
+
+def topic_record_subject_name_strategy(ctx, record_name: Optional[str]) -> Optional[str]:
+    """
+    Constructs a subject name in the form of {topic}-{record_name}.
+
+    Args:
+        ctx (SerializationContext): Metadata pertaining to the serialization
+            operation.
+
+        record_name (Optional[str]): Record name.
+
+    """
+    return ctx.topic + "-" + record_name if record_name is not None else None
+
+
+def record_subject_name_strategy(ctx, record_name: Optional[str]) -> Optional[str]:
+    """
+    Constructs a subject name in the form of {record_name}.
+
+    Args:
+        ctx (SerializationContext): Metadata pertaining to the serialization
+            operation.
+
+        record_name (Optional[str]): Record name.
+
+    """
+    return record_name if record_name is not None else None
+
+
+def reference_subject_name_strategy(ctx, schema_ref: SchemaReference) -> Optional[str]:
+    """
+    Constructs a subject reference name in the form of {reference name}.
+
+    Args:
+        ctx (SerializationContext): Metadata pertaining to the serialization
+            operation.
+
+        schema_ref (SchemaReference): SchemaReference instance.
+
+    """
+    return schema_ref.name if schema_ref is not None else None
