+++ src/confluent_kafka/schema_registry/_sync/protobuf.py
@@ -26,9 +26,10 @@ from google.protobuf.descriptor import Descriptor, FileDescriptor
 from google.protobuf.message import DecodeError, Message
 from google.protobuf.message_factory import GetMessageClass
 
-from confluent_kafka.schema_registry.common import (_MAGIC_BYTE, _ContextStringIO,
+from confluent_kafka.schema_registry import (_MAGIC_BYTE,
                reference_subject_name_strategy,
                topic_subject_name_strategy)
+from confluent_kafka.schema_registry.common.protobuf import _ContextStringIO
 from confluent_kafka.schema_registry.schema_registry_client import SchemaRegistryClient
 from confluent_kafka.schema_registry.common.protobuf import _bytes, _create_index_array, _init_pool, _is_builtin, _schema_to_str, _str_to_proto, transform
 from confluent_kafka.schema_registry.rule_registry import RuleRegistry
