+++ src/confluent_kafka/schema_registry/common/__init__.py
@@ -19,73 +19,3 @@ from io import BytesIO
 from typing import Optional
 
 from .schema_registry_client import SchemaReference
-
-_MAGIC_BYTE = 0
-
-def topic_subject_name_strategy(ctx, record_name: Optional[str]) -> Optional[str]:
-    """
-    Constructs a subject name in the form of {topic}-key|value.
-
-    Args:
-        ctx (SerializationContext): Metadata pertaining to the serialization
-            operation.
-
-        record_name (Optional[str]): Record name.
-
-    """
-    return ctx.topic + "-" + ctx.field
-
-
-def topic_record_subject_name_strategy(ctx, record_name: Optional[str]) -> Optional[str]:
-    """
-    Constructs a subject name in the form of {topic}-{record_name}.
-
-    Args:
-        ctx (SerializationContext): Metadata pertaining to the serialization
-            operation.
-
-        record_name (Optional[str]): Record name.
-
-    """
-    return ctx.topic + "-" + record_name if record_name is not None else None
-
-
-def record_subject_name_strategy(ctx, record_name: Optional[str]) -> Optional[str]:
-    """
-    Constructs a subject name in the form of {record_name}.
-
-    Args:
-        ctx (SerializationContext): Metadata pertaining to the serialization
-            operation.
-
-        record_name (Optional[str]): Record name.
-
-    """
-    return record_name if record_name is not None else None
-
-
-def reference_subject_name_strategy(ctx, schema_ref: SchemaReference) -> Optional[str]:
-    """
-    Constructs a subject reference name in the form of {reference name}.
-
-    Args:
-        ctx (SerializationContext): Metadata pertaining to the serialization
-            operation.
-
-        schema_ref (SchemaReference): SchemaReference instance.
-
-    """
-    return schema_ref.name if schema_ref is not None else None
-
-
-class _ContextStringIO(BytesIO):
-    """
-    Wrapper to allow use of StringIO via 'with' constructs.
-    """
-
-    def __enter__(self):
-        return self
-
-    def __exit__(self, *args):
-        self.close()
-        return False
